+++ src/classes/managers/orders/errors/modify-position.error.ts
@@ -1,0 +1,9 @@
+import { cTraderXError } from '../../../models/ctrader-x-error.model';
+
+export class ModifyPositionError extends cTraderXError {
+    constructor(error: unknown) {
+        super(
+            `Modify position error: ${cTraderXError.getMessageError(error)}`,
+        );
+    }
+}